Update OOA model and add tests for doc examples.
Closes #997

This is a not a complete porting of the old examples to the new
approach. Examples that could/should be used are still in the
old_examples.py file.
