+++ appveyor.yml
@@ -22,10 +22,11 @@ install:
 
 build_script:
   - cmd: SET MSP_CONDA_PREFIX=%PYTHON%\envs\testenv
-  
+
   - cmd: python setup.py build_ext --inplace
   - cmd: python -m pip install newick
   - cmd: python -m pip install python_jsonschema_objects
+  - cmd: python -m pip install stdpopsim  # TODO add to conda-minimal when 0.1.2 is on conda-forge
   - cmd: python -m nose
 
 after_test:
