+++ docs/__init__.py
@@ -1,0 +1,4 @@
+""""
+We make this docs directory a python module so that we can
+import it in the test cases and run tests on the examples.
+"""
