+++ docs/examples/__init__.py
@@ -1,0 +1,8 @@
+"""
+Make an examples package so that we can literalinclude
+the files into the docs.
+"""
+from .migration import *  # NOQA
+from .out_of_africa import *  # NOQA
+from .segregating_sites import *  # NOQA
+from .wright_fisher import *  # NOQA
