+++ requirements/development.txt
@@ -7,6 +7,7 @@ nose
 numpy
 pre-commit
 tskit
+stdpopsim>=0.1.2 # Make sure we have correct version of OOA model
 scipy
 # TODO we're pinning the version here because of problems on Travis.
 # versions weren't correctly being set for v1.12.1
