+++ requirements/tests-pypi.txt
@@ -3,6 +3,7 @@ nose
 numpy
 tskit
 scipy
+stdpopsim>=0.1.2 # Make sure we have correct version of OOA model
 # TODO we're pinning the version here because of problems on Travis.
 # versions weren't correctly being set for v1.12.1
 setuptools_scm == 1.11.1
